+++ trunk/FinalTwinkie/FinalTwinkie/GameObjects/Building.cpp
@@ -53,11 +53,11 @@ void CBuilding::Render(void)
 		//Health
 		RECT rect;
 		rect.top=(LONG)360; rect.left=(LONG)11; rect.right=(LONG)177; rect.bottom=(LONG)381;
-		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (GetPosX()-41)+pCam->GetPosX(), (GetPosY()-80)+pCam->GetPosY(), 0.5f,0.5f,&rect);
+		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (int)((GetPosX()-41)+pCam->GetPosX()), (int)((GetPosY()-80)+pCam->GetPosY()), 0.5f,0.5f,&rect);
 		rect.top=(LONG)364; rect.left=(LONG)194; rect.right=(LONG)(194+((353-194)*(GetHealth()/GetMaxHealth()))); rect.bottom=(LONG)378;
-		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (GetPosX()-41)+pCam->GetPosX(), (GetPosY()-80)+pCam->GetPosY(), 0.5f,0.5f,&rect);
+		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (int)((GetPosX()-41)+pCam->GetPosX()), (int)((GetPosY()-80)+pCam->GetPosY()), 0.5f,0.5f,&rect);
 		rect.top=(LONG)391; rect.left=(LONG)351; rect.right=(LONG)(351+((509-351)*(GetHealth()/GetMaxHealth()))); rect.bottom=(LONG)403;
-		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (GetPosX()-41)+pCam->GetPosX(), (GetPosY()-80)+pCam->GetPosY(), 0.5f,0.5f,&rect);
+		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (int)((GetPosX()-41)+pCam->GetPosX()), (int)((GetPosY()-80)+pCam->GetPosY()), 0.5f,0.5f,&rect);
 	}
 
 	if(GetPosX()+pCam->GetPosX() >= -100 && GetPosX()+pCam->GetPosX() <= CGame::GetInstance()->GetWidth()+100 && GetPosY()+pCam->GetPosY() >= -100 && GetPosY()+pCam->GetPosY() <= CGame::GetInstance()->GetHeight()+100)
@@ -149,9 +149,9 @@ CBuilding::~CBuilding(void)
 void CBuilding::TakeDamage(int nDamage)
 {
 
-	int m_fHealth = GetHealth();
+	float m_fHealth = GetHealth();
 	m_fHealth -= nDamage;
-	SetHealth(m_fHealth);
+	SetHealth((float)m_fHealth);
 	if(m_fHealth<0)
 	{
 		m_fHealth=0;
