+++ trunk/FinalTwinkie/FinalTwinkie/GameObjects/Factory.cpp
@@ -25,18 +25,18 @@ void Factory::Render(void)
 	Camera* pCam=Camera::GetInstance();
 
 	Camera* C=Camera::GetInstance();
-	CSGD_TextureManager::GetInstance()->Draw(GetImageID(), (GetPosX()+pCam->GetPosX())-GetWidth()/2, (GetPosY()+pCam->GetPosY())-GetHeight()/2);
+	CSGD_TextureManager::GetInstance()->Draw(GetImageID(), (int)((GetPosX()+pCam->GetPosX())-GetWidth()/2), (int)((GetPosY()+pCam->GetPosY())-GetHeight()/2));
 
 	if(GetHealth() < GetMaxHealth())
 	{
 		//Health
 		RECT rect;
 		rect.top=(LONG)360; rect.left=(LONG)11; rect.right=(LONG)177; rect.bottom=(LONG)381;
-		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (GetPosX()-41)+pCam->GetPosX(), (GetPosY()-40)+pCam->GetPosY(), 0.5f,0.5f,&rect);
+		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (int)((GetPosX()-41)+pCam->GetPosX()), (int)((GetPosY()-40)+pCam->GetPosY()), 0.5f,0.5f,&rect);
 		rect.top=(LONG)364; rect.left=(LONG)194; rect.right=(LONG)(194+((353-194)*(GetHealth()/GetMaxHealth()))); rect.bottom=(LONG)378;
-		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (GetPosX()-41)+pCam->GetPosX(), (GetPosY()-40)+pCam->GetPosY(), 0.5f,0.5f,&rect);
+		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (int)((GetPosX()-41)+pCam->GetPosX()), (int)((GetPosY()-40)+pCam->GetPosY()), 0.5f,0.5f,&rect);
 		rect.top=(LONG)391; rect.left=(LONG)351; rect.right=(LONG)(351+((509-351)*(GetHealth()/GetMaxHealth()))); rect.bottom=(LONG)403;
-		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (GetPosX()-41)+pCam->GetPosX(), (GetPosY()-40)+pCam->GetPosY(), 0.5f,0.5f,&rect);
+		CSGD_TextureManager::GetInstance()->Draw(m_nHPID, (int)((GetPosX()-41)+pCam->GetPosX()), (int)((GetPosY()-40)+pCam->GetPosY()), 0.5f,0.5f,&rect);
 	}
 }
 
@@ -44,10 +44,10 @@ RECT Factory::GetRect(void)
 {
 	Camera* C=Camera::GetInstance();
 	RECT rect={0,0,0,0};
-	rect.top=GetPosY()+C->GetPosY()-GetHeight()/2;
-	rect.bottom=GetPosY()+C->GetPosY()+GetHeight()/2;
-	rect.left=GetPosX()+C->GetPosX()-GetWidth()/2;
-	rect.right=GetPosX()+C->GetPosX()+GetWidth()/2;
+	rect.top=(LONG)(GetPosY()+C->GetPosY()-GetHeight()/2);
+	rect.bottom=(LONG)(GetPosY()+C->GetPosY()+GetHeight()/2);
+	rect.left=(LONG)(GetPosX()+C->GetPosX()-GetWidth()/2);
+	rect.right=(LONG)(GetPosX()+C->GetPosX()+GetWidth()/2);
 	return rect;
 }
